second/requirements.txt
flask==2.2.3
psycopg2-binary==2.9.5
python-dotenv==1.0.0
typing==3.7.4.3
